# openapi.yaml
openapi: 3.0.0
info:
  title: TradeOn API
  version: 1.0.0
  description: 'TradeOn은 사용자 간 물물교환 기반의 중고 거래 플랫폼입니다.


    이 API 문서는 다음과 같은 기능들을 포함합니다:

    - 사용자 회원가입, 로그인 및 정보 조회

    - 교환 요청 및 거래 이력 관리

    - 사용자 리뷰 시스템

    - 가치 평가 게시판

    - 하이브리드 검색기능

    - 게시물 등록/삭제/조회

    - 채팅 및 거래 약속

    - 거래 데이터 분석 및 시각화


    JWT 인증이 필요한 API의 경우 Authorize 버튼을 통해 토큰을 입력해주세요.'
servers:
- url: https://673e-121-140-48-141.ngrok-free.app
  description: 로컬
...
paths:
...
  /api/v1/valuations:
    get:
      tags:
        - Valuation
      summary: 가치 평가 게시글 목록 조회
      description: 인증된 사용자의 가치 평가 게시글을 페이지네이션 및 카테고리 필터를 통해 최신순으로 조회합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: 페이지 번호 (기본값 1)
          required: false
          schema:
            type: integer
            example: 1
        - name: per_page
          in: query
          description: 페이지당 게시글 수 (기본값 10)
          required: false
          schema:
            type: integer
            example: 10
        - name: category
          in: query
          description: 카테고리 필터 (예시-> 전자기기, 패션 등)
          required: false
          schema:
            type: string
            example: 전자기기
      responses:
        "200":
          description: 게시글 목록 반환
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValuationPostListResponse"
        "401":
          description: 인증 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


  
    post:
      tags:
        - Valuation
      summary: 가치 평가 게시글 작성
      description: 인증된 사용자가 새 가치 평가 게시글을 등록합니다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValuationPostCreateRequest"
      responses:
        "201":
          description: 게시글 생성 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValuationPostDetail"
        "400":
          description: 필수 입력 항목 누락 또는 유효성 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ──[3] GET /api/v1/valuations/{postid} ──────────────────────────────────────────────
  /api/v1/valuations/{postid}:
    get:
      tags:
        - Valuation
      summary: 가치 평가 게시글 상세 조회
      description: 특정 게시글의 상세 정보를 조회합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: postid
          in: path
          description: 조회할 게시글 ID
          required: true
          schema:
            type: integer
            example: 123
      responses:
        "200":
          description: 게시글 상세 정보 반환
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValuationPostDetail"
        "401":
          description: 인증 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 게시글을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - Valuation
      summary: 가치 평가 게시글 삭제
      description: 작성자만 자신의 가치 평가 게시글을 삭제할 수 있습니다.
      security:
        - bearerAuth: []
      parameters:
        - name: postid
          in: path
          description: 삭제할 게시글 ID
          required: true
          schema:
            type: integer
            example: 123
      responses:
        "200":
          description: 게시글 삭제 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessageResponse"
        "401":
          description: 인증 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 권한 없음 (작성자 아님)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 게시글을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ──[4] POST /api/v1/valuations/{postid}/price ─────────────────────────────────────
  /api/v1/valuations/{postid}/price:
    post:
      tags:
        - Valuation
      summary: 평가 의견 등록
      description: 특정 게시글에 대한 가격 의견을 등록합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: postid
          in: path
          description: 평가할 게시글 ID
          required: true
          schema:
            type: integer
            example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValuationOpinionCreateRequest"
      responses:
        "201":
          description: 의견 등록 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessageResponse"
        "400":
          description: 필수 입력 항목 누락 또는 유효성 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 게시글을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ──[5] GET /api/v1/valuations/{postid}/average ────────────────────────────────────
  /api/v1/valuations/{postid}/average:
    get:
      tags:
        - Valuation
      summary: 게시글의 평균 가격 조회
      description: 특정 가치 평가 게시글에 대한 유효한 의견들의 평균 가격과 개수를 반환합니다. JWT 인증 필요.
      security:
        - bearerAuth: []
      parameters:
        - name: postid
          in: path
          description: 평균 가격을 조회할 게시글 ID
          required: true
          schema:
            type: integer
            example: 123
      responses:
        "200":
          description: 평균 가격 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValuationAverageResponse"
        "401":
          description: 인증 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 게시글을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ValuationPostListResponse:
      type: object
      properties:
        total:
          type: integer
          description: 전체 게시글 수
          example: 25
        page:
          type: integer
          description: 현재 페이지 번호
          example: 1
        per_page:
          type: integer
          description: 페이지당 게시글 수
          example: 10
        posts:
          type: array
          description: 게시글 요약 리스트
          items:
            $ref: "#/components/schemas/ValuationPostSummary"
      required:
        - total
        - page
        - per_page
        - posts

    # ──[B] 게시글 요약 스키마 (리스트 아이템)──────────────────────────────────
    ValuationPostSummary:
      type: object
      properties:
        postId:
          type: integer
          description: 게시글 ID
          example: 123
        title:
          type: string
          description: 게시글 제목
          example: "노트북 중고 팝니다"
        averagePrice:
          type: number
          format: float
          description: 현재까지 유효한 의견들의 평균 가격
          example: 500000.00
        createdAt:
          type: string
          format: date-time
          description: 게시글 생성 시각 (YYYY-MM-DD HH:MM)
          example: "2025-06-01 14:30"
        category:
          type: string
          description: 카테고리
          example: "전자기기"
        image:
          type: string
          description: 게시글 대표 이미지 URL (선택)
          example: https://example.com/images/123.jpg
        totalEvaluations:
          type: integer
          description: 유효한 평가(의견) 개수
          example: 5
      required:
        - postId
        - title
        - averagePrice
        - createdAt
        - category
        - image


    # ──[C] 게시글 상세 스키마────────────────────────────────────────────────
    ValuationPostDetail:
      type: object
      properties:
        postId:
          type: integer
          description: 게시글 ID
          example: 123
        title:
          type: string
          description: 게시글 제목
          example: "노트북 중고 팝니다"
        description:
          type: string
          description: 게시글 상세 설명
          example: "최상급 컨디션의 노트북을 저렴하게 팝니다."
        category:
          type: string
          description: 카테고리
          example: "전자기기"
        averagePrice:
          type: number
          format: float
          description: 현재까지 유효한 의견들의 평균 가격
          example: 500000.00
        totalEvaluations:
          type: integer
          description: 유효한 평가(의견) 개수
          example: 5
        createdAt:
          type: string
          format: date-time
          description: 게시글 생성 시각 (YYYY-MM-DD HH:MM)
          example: "2025-06-01 14:30"
        image:
          type: string
          description: 게시글 대표 이미지 URL (선택)
          example: https://example.com/images/123.jpg
      required:
        - postId
        - title
        - description
        - category
        - averagePrice
        - totalEvaluations
        - createdAt
        - image

    # ──[D] 게시글 생성 요청 스키마────────────────────────────────────────────
    ValuationPostCreateRequest:
      type: object
      properties:
        title:
          type: string
          description: 게시글 제목
          example: "노트북 중고 팝니다"
        description:
          type: string
          description: 게시글 상세 설명
          example: "최상급 컨디션의 노트북을 저렴하게 팝니다."
        category:
          type: string
          description: 카테고리
          example: "전자기기"
        image:
          type: string
          description: 게시글 대표 이미지 URL (선택)
          example: https://example.com/images/123.jpg
      required:
        - title
        - description
        - category

    # ──[E] 의견 등록 요청 스키마────────────────────────────────────────────
    ValuationOpinionCreateRequest:
      type: object
      properties:
        price:
          type: integer
          description: 게시글에 대한 의견 가격
          example: 480000
      required:
        - price

    # ──[F] 평균 가격 조회 응답 스키마──────────────────────────────────────────
    ValuationAverageResponse:
      type: object
      properties:
        postId:
          type: integer
          description: 게시글 ID
          example: 123
        validCount:
          type: integer
          description: 유효한 의견 개수
          example: 5
        averagePrice:
          type: number
          format: float
          description: 유효한 의견들의 평균 가격
          example: 500000.00
      required:
        - postId
        - validCount
        - averagePrice

    # ──[G] 성공 메시지 공통 응답──────────────────────────────────────────────
    SuccessMessageResponse:
      type: object
      properties:
        msg:
          type: string
          description: 성공 메시지
          example: "게시글 삭제 성공"
      required:
        - msg
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        msg:
          type: string
          example: 오류 메시지입니다.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []